+++ src/inject/dynamic-theme/inline-style.ts
@@ -309,7 +309,50 @@ function shouldIgnoreInlineStyle(element: HTMLElement, selectors: string[]): boo
     return false;
 }
 
+const LOOP_DETECTION_THRESHOLD = 1000;
+const MAX_LOOP_CYCLES = 10;
+const elementsLastChanges = new WeakMap<Node, number>();
+const elementsLoopCycles = new WeakMap<Node, number>();
+
+const SMALL_SVG_THRESHOLD = 32;
+const svgNodesRoots = new WeakMap<Node, SVGSVGElement | null>();
+const svgRootSizeTestResults = new WeakMap<SVGSVGElement, boolean>();
+
+function getSVGElementRoot(svgElement: SVGElement): SVGSVGElement | null {
+    if (!svgElement) {
+        return null;
+    }
+
+    if (svgNodesRoots.has(svgElement)) {
+        return svgNodesRoots.get(svgElement)!
+    }
+
+    if (svgElement instanceof SVGSVGElement) {
+        return svgElement;
+    }
+
+    const parent = svgElement.parentNode as SVGElement;
+    const root = getSVGElementRoot(parent!);
+    svgNodesRoots.set(svgElement, root);
+    return root;
+}
+
 export function overrideInlineStyle(element: HTMLElement, theme: Theme, ignoreInlineSelectors: string[], ignoreImageSelectors: string[]): void {
+    if (elementsLastChanges.has(element)) {
+        if (Date.now() - elementsLastChanges.get(element)! < LOOP_DETECTION_THRESHOLD) {
+            const cycles = elementsLoopCycles.get(element) ?? 0;
+            elementsLoopCycles.set(element, cycles + 1);
+        }
+        if ((elementsLoopCycles.get(element) ?? 0) >= MAX_LOOP_CYCLES) {
+            return;
+        }
+    }
+
+    // ProseMirror editor rebuilds entire HTML after style changes
+    if (element.parentElement?.dataset.nodeViewContent) {
+        return;
+    }
+
     const cacheKey = getInlineStyleCacheKey(element, theme);
     if (cacheKey === inlineStyleCache.get(element)) {
         return;
@@ -451,7 +494,6 @@ export function overrideInlineStyle(element: HTMLElement, theme: Theme, ignoreIn
 
     if (isSVGElement) {
         if (element.hasAttribute('fill')) {
-            const SMALL_SVG_LIMIT = 32;
             const value = element.getAttribute('fill')!;
             if (value !== 'none') {
                 if (!(element instanceof SVGTextElement)) {
@@ -460,8 +502,25 @@ export function overrideInlineStyle(element: HTMLElement, theme: Theme, ignoreIn
                     // and it will cause a layout of unstyled content which results in white flashes.
                     // Therefore, check if the DOM is at the `complete` readystate.
                     const handleSVGElement = () => {
-                        const {width, height} = element.getBoundingClientRect();
-                        const isBg = (width > SMALL_SVG_LIMIT || height > SMALL_SVG_LIMIT);
+                        let isSVGSmall = false;
+                        const root = getSVGElementRoot(element);
+                        if (!root) {
+                            return;
+                        }
+                        if (svgRootSizeTestResults.has(root)) {
+                            isSVGSmall = svgRootSizeTestResults.get(root)!;
+                        } else {
+                            const svgBounds = root.getBoundingClientRect();
+                            isSVGSmall = svgBounds.width * svgBounds.height <= Math.pow(SMALL_SVG_THRESHOLD, 2);
+                            svgRootSizeTestResults.set(root, isSVGSmall);
+                        }
+                        let isBg: boolean;
+                        if (isSVGSmall) {
+                            isBg = false;
+                        } else {
+                            const {width, height} = element.getBoundingClientRect();
+                            isBg = (width > SMALL_SVG_THRESHOLD || height > SMALL_SVG_THRESHOLD);
+                        }
                         setCustomProp('fill', isBg ? 'background-color' : 'color', value);
                     };
 
@@ -522,4 +581,6 @@ export function overrideInlineStyle(element: HTMLElement, theme: Theme, ignoreIn
         element.removeAttribute(overrides[cssProp].dataAttr);
     });
     inlineStyleCache.set(element, getInlineStyleCacheKey(element, theme));
+
+    elementsLastChanges.set(element, Date.now());
 }