+++ src/inject/dynamic-theme/stylesheet-modifier.ts
@@ -51,6 +51,9 @@ export function createStyleSheetModifier(): StyleSheetModifier {
         if (isMediaRule(rule.parentRule)) {
             cssText = `${rule.parentRule.media.mediaText} { ${cssText} }`;
         }
+        if (isLayerRule(rule.parentRule)) {
+            cssText = `${rule.parentRule.name} { ${cssText} }`;
+        }
         return getHashCode(cssText);
     }
 
@@ -169,6 +172,11 @@ export function createStyleSheetModifier(): StyleSheetModifier {
             const viewTransitionSelector = selector.includes('::view-transition-');
             if (emptyIsWhereSelector || viewTransitionSelector) {
                 selectorText = '.darkreader-unsupported-selector';
+            }
+            // ::picker(select) becomes ::picker,
+            // but cannot be parsed later (Chrome bug)
+            if (isChromium && selectorText.endsWith('::picker')) {
+                selectorText = selectorText.replaceAll('::picker', '::picker(select)');
             }
 
             let ruleText = `${selectorText} {`;