+++ src/stubs/popup/plus-body.tsx
@@ -7,3 +7,11 @@ export function PlusBody(): Malevic.Child {
 export function activate(_email: string, _key: string): Promise<boolean> {
     return new Promise((resolve) => resolve(false));
 }
+
+export function shouldUseAltUI() {
+    return false;
+}
+
+export function initAltUI(): Promise<void> {
+    return new Promise((resolve) => resolve());
+}