+++ src/utils/state-manager.ts
@@ -13,7 +13,7 @@ export class StateManager<T extends Record<string, unknown>> {
     constructor(localStorageKey: string, parent: any, defaults: T, logWarn: (log: string) => void){
         if (isNonPersistent) {
             function addListener(listener: (data: T) => void) {
-                chrome.storage.local.onChanged.addListener((changes) => {
+                chrome.storage.local.onChanged.addListener((changes: Record<string, any>) => {
                     if (localStorageKey in changes) {
                         listener(changes[localStorageKey].newValue);
                     }